Add guard on block/treeview buttons in ToolBar

+++ src/client/src/components/toolbar/handlers/OnToolbarClick.ts
@@ -10,28 +10,30 @@ import { TextResources } from "../../../assets/text/TextResources";
 import { removeSecondaryNode } from "../../../redux/store/secondaryNode/actions";
 import { removeActiveBlockNode, removeActiveEdge, removeActiveNode } from "../../../redux/store/project/actions";
 
-export const OnElectroClick = (dispatch: Dispatch) => {
-  dispatch(toggleElectroView());
-};
-
-export const OnFilterClick = (dispatch: Dispatch, open: boolean) => {
-  dispatch(setFilterMenuVisibility(!open));
-};
+export const OnElectroClick = (dispatch: Dispatch) => dispatch(toggleElectroView());
+export const OnFilterClick = (dispatch: Dispatch, open: boolean) => dispatch(setFilterMenuVisibility(!open));
 
 export const OnFitToScreenClick = (isTreeView: boolean, viewportData: ViewportData, secondaryNode: Node) => {
   if (isTreeView) return;
   SetFitToScreen(viewportData, secondaryNode !== null);
 };
 
-export const OnBlockViewClick = (numberOfSelectedElements: number, viewportData: ViewportData, dispatch: Dispatch) => {
-  if (!ValidateBlockViewClick(numberOfSelectedElements, dispatch)) return;
+export const OnBlockViewClick = (
+  numberOfSelectedElements: number,
+  viewportData: ViewportData,
+  isTreeView: boolean,
+  dispatch: Dispatch
+) => {
+  if (!isTreeView || !ValidateBlockViewClick(numberOfSelectedElements, dispatch)) return;
 
   SetFitToScreen(viewportData, false);
   dispatch(removeSecondaryNode());
   dispatch(changeFlowView(VIEW_TYPE.BLOCKVIEW as ViewType));
 };
 
-export const OnTreeViewClick = (setSelectedNodes: (nodeIds: string[]) => void, dispatch: Dispatch) => {
+export const OnTreeViewClick = (setSelectedNodes: (nodeIds: string[]) => void, isTreeView: boolean, dispatch: Dispatch) => {
+  if (isTreeView) return;
+
   // When opening TreeView all selectedItems are removed
   setSelectedNodes([]);
   dispatch(removeActiveNode());
